# PropertyValueConverter improvements