--- a/src/Serilog/Core/IScalarConversionPolicy.cs
+++ b/src/Serilog/Core/IScalarConversionPolicy.cs
@@ -12,9 +12,8 @@ namespace Serilog.Core
         /// If supported, convert the provided value into an immutable scalar.
         /// </summary>
         /// <param name="value">The value to convert.</param>
-        /// <param name="propertyValueFactory">Recursively apply policies to convert additional values.</param>
         /// <param name="result">The converted value, or null.</param>
         /// <returns>True if the value could be converted under this policy.</returns>
-        bool TryConvertToScalar(object value, ILogEventPropertyValueFactory propertyValueFactory, out ScalarValue result);
+        bool TryConvertToScalar(object value, out ScalarValue result);
     }
 }