--- a/src/Serilog/Policies/ByteArrayScalarConversionPolicy.cs
+++ b/src/Serilog/Policies/ByteArrayScalarConversionPolicy.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2013-2015 Serilog Contributors
+﻿// Copyright 2013-2017 Serilog Contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -25,7 +25,7 @@ namespace Serilog.Policies
     {
         const int MaximumByteArrayLength = 1024;
 
-        public bool TryConvertToScalar(object value, ILogEventPropertyValueFactory propertyValueFactory, out ScalarValue result)
+        public bool TryConvertToScalar(object value, out ScalarValue result)
         {
             var bytes = value as byte[];
             if (bytes == null)