--- a/test/Serilog.PerformanceTests/xunit.runner.json
+++ b/test/Serilog.PerformanceTests/xunit.runner.json
@@ -1,0 +1,3 @@
+﻿{
+  "shadowCopy": false
+}