… results/netcoreapp1.1/AllocationsBenchmark-report-github.md
… src/Serilog/Policies/NullableScalarConversionPolicy.cs
… test/Serilog.Tests/Parameters/PropertyValueConverterTests.cs