@nblumhardt 2017-06-06
Great that the memory diagnoser is working now; I'll give that a shot on #977 👍 

LGTM, merging.

@merbla it looks like the new perf test build on AppVeyor has clobbered the regular CI build... Should we just disable it for now? Cheers!