@merbla 2017-06-06
@nblumhardt I have disabled that build.